[DevOverlay] Add CodeFrame
