--- packages/next/src/client/components/react-dev-overlay/_experimental/internal/components/CodeFrame/index.tsx
@@ -1,1 +1,0 @@
-export { CodeFrame } from './CodeFrame'
