--- packages/next/src/client/components/react-dev-overlay/_experimental/internal/components/CodeFrame/styles.tsx
@@ -1,53 +1,0 @@
-import { noop as css } from '../../helpers/noop-template'
-
-const styles = css`
-  [data-nextjs-codeframe] {
-    overflow: auto;
-    border-radius: var(--size-gap-half);
-    background-color: var(--color-ansi-bg);
-    color: var(--color-ansi-fg);
-    margin-bottom: var(--size-gap-double);
-  }
-  [data-nextjs-codeframe]::selection,
-  [data-nextjs-codeframe] *::selection {
-    background-color: var(--color-ansi-selection);
-  }
-  [data-nextjs-codeframe] * {
-    color: inherit;
-    background-color: transparent;
-    font-family: var(--font-stack-monospace);
-  }
-
-  [data-nextjs-codeframe] > * {
-    margin: 0;
-    padding: calc(var(--size-gap) + var(--size-gap-half))
-      calc(var(--size-gap-double) + var(--size-gap-half));
-  }
-  [data-nextjs-codeframe] > div {
-    display: inline-block;
-    width: auto;
-    min-width: 100%;
-    border-bottom: 1px solid var(--color-ansi-bright-black);
-  }
-  [data-nextjs-codeframe] > div > p {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    cursor: pointer;
-    margin: 0;
-  }
-  [data-nextjs-codeframe] > div > p:hover {
-    text-decoration: underline dotted;
-  }
-  [data-nextjs-codeframe] div > p > svg {
-    width: auto;
-    height: 1em;
-    margin-left: 8px;
-  }
-  [data-nextjs-codeframe] div > pre {
-    overflow: hidden;
-    display: inline-block;
-  }
-`
-
-export { styles }
